+++ src/components/choropleth2/legenda/hooks/useSafetyRegionLegendaData.ts
@@ -1,30 +1,6 @@
-import {
-  ChoroplethThresholds,
-  TRegionMetricName,
-  TRegionsNursingHomeMetricName,
-} from '../../shared';
-
+import { TRegionMetricName, TRegionsNursingHomeMetricName } from '../../shared';
+import { getSelectedThreshold } from '../utils';
 import { useLegendaItems } from './useLegendaItems';
-import { regionThresholds } from '~/components/choropleth2/region-thresholds';
-
-export function getSelectedThreshold(
-  metricName?: TRegionMetricName,
-  metricValueName?: string
-) {
-  if (!metricName) {
-    return;
-  }
-
-  // Even if a metricValueName is passed in, there's not necessarily
-  // a threshold defined for this. In that case we fall back to the threshold
-  // that exists for the metric name.
-  const thresholdInfo =
-    (metricValueName
-      ? (regionThresholds as any)?.[metricName]?.[metricValueName]
-      : regionThresholds[metricName]) ?? regionThresholds[metricName];
-
-  return thresholdInfo as ChoroplethThresholds;
-}
 
 export function useSafetyRegionLegendaData(
   metric: TRegionMetricName,
